♻️ [refactor] 리뷰 코멘트 적용 #10
